+++ src/main/java/com/cju/shoppingmall/product/controller/ProductController.java
@@ -1,4 +1,4 @@
-package com.cju.shoppingmall.controller;
+package com.cju.shoppingmall.product.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
