+++ src/main/java/com/cju/shoppingmall/product/entity/CategoryDto.java
@@ -1,0 +1,11 @@
+package com.cju.shoppingmall.product.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CategoryDto {
+    private Long id;
+    private String name;
+}
