+++ src/main/java/com/cju/shoppingmall/product/repository/CategoryRepository.java
@@ -4,8 +4,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cju.shoppingmall.product.entity.Category;
 
+import java.util.List;
 import java.util.Optional;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
     Optional<Category> findByName(String name);
+
+    List<Category> findByParentIsNullAndIsActiveTrue();
+    List<Category> findByParentIdAndIsActiveTrue(Long parentId);
 }
