+++ src/main/java/com/cju/shoppingmall/product/service/ProductService.java
@@ -2,7 +2,7 @@ package com.cju.shoppingmall.product.service;
 
 import java.util.List;
 
-import com.cju.shoppingmall.controller.ProductRegisterForm;
+import com.cju.shoppingmall.product.controller.ProductRegisterForm;
 import com.cju.shoppingmall.member.entity.Member;
 import com.cju.shoppingmall.product.entity.Category;
 import com.cju.shoppingmall.product.entity.Product;
